Work-in-progress:
KNXConnectionManager is implemented to manage KNX connections which may be severely limited in numbers (down to 
one available connection in case of many KNX IP gateways).

Current implementation is 'aggressive' in that it attempts to maintain one persistent connection to the IP 
gateway. However, since the KNX event system essentially delegates all operations to the connection manager, it 
is trivial to implement other connection policies depending to deployment needs.