added SensorLinker.java to store the UI property for a sensor. added SelectColorWindow.java to provide a color palette. Modified SensorServiceImpl.java to let avoid lazy initialization so that we can get the state information for a customer sensor. Modified LabelPropertyForm.java to add Sensor UI property edit function. Modified ImagePropertyForm.java to add Sensor UI property edit function.
Modified UIComponent.java to provide a build method to get a a copy of a UIComponent and a method to create a new UIComponent.
Modified UIImage.java, UILabel.java to let them have a SensorLinker to record the property for the sensor they have. 

update the xml template for panel.xml and controller.xml.