MODELER-149 	 Horizontal/Vertical slider UI is incorrect by default/customize Added VerticalScreenSlider,java : a class for a vertical slider. Added HorizontalScreenSlider.java: a class for a horizontal slider.
removed FlexSliderTable.java: remove the previous UI for a screen slider. 

Modified TreePanelBuilder.java to add a vertical slider.
Modified ScreenSlider.java to update the UI for a slider. 
Modified Icons.java to add a icon for a vertical slider. 
Modified SliderPropertyForm : remove vertical & horizontal setting for a slider. 

update the default icons for a slider .