isControllerXMLChanged method added to replace the current method in XMLListenSharingData.
Spring sourced reference to GatewayManagerService added so gateways can be restarted when an XML change is detected, old version had lots of code for restarting the status cache which duplicates the code used to start the status cache in the first place.

A restart method has been added to the GatewayManagerService which can be called from here, this then makes use of the same code used to first start the gateways and initialise the status cache.