re ANDROID-85 clean up client-side controller fail-over mechanism
Add the beginning of a controller communications abstraction, in
the form of the ControllerService interface, along with an
implementation for communicating with a controller via HTTP or
HTTPS using the OpenRemote 2.0 Controller XML REST API.

Ideally, controller failover would be attempted in an implementation
of this interface.

This should also help with testing, as making an implementation
of the ControllerService interface that mocks a real controller
can be done very easily, eliminating the need to communicate
with a real controller over the network in JUnit tests under
most circumstances.