Add deployer bean into the configuration that centralizes a lot of the XML mapping and object model functionality.
Make dependencies explicit (instead of auto-wire) on 
two services that are deprecated and will be removed
(mostly replaced by deployer): PollingMachineService
and ControllerXMLChangeService

SensorBuilder is the first XML-to-Object builder that
works directly through the deployment service (the goal
is to allow multiple object builders to co-exist as
the schema evolves and migration needs to be supported).

Add explicit dependencies to the other *Builder beans
that do not yet work through the new deployment service.