Rewrite the sensor retrieval logic.
This avoids code duplication which was causing an issue earlier
(a bug was fixed in one part of the code but not in the 
 copy + pasted code). So no more code duplication to prevent
similar issues occurring later.

Also tried to figure out the logic in the code itself and
added some documentation of it. There's still clearly something
about this part that doesn't smell right, or there are
non-obvious side-effects that are not clear in the undocumented
code. Won't touch further for now since there aren't unit
tests to back this up.