Work In Progress: Gateway and Protocol Implementations added.
The Gateway is the main component of the proposed system and can only be instantiated and used by a GatewayManager (yet to be implemented). The GatewayManager will send command requests to the Gateway which are then dealt with asynchronously.

The Gateway manages two internal threads: -

The gatewayThread which deals with actioning the incoming command requests, polling sensor commands that belong to that gateway and handling responses received from the protocol implementation.

The protocolThread which is responsible for invoking the send() and read() Methods of the protocol implementation, it also deals with listening to messages from the physical device if the protocol implements the ActiveProtocol interface.