This is a DREAMBOX-ONLY patch to deal with an issue with DreamBox HTTP response XML documents where the HTTP charset does *NOT* match the charset claimed in the returned XML document's <?xml?> directive (XML claims to be in UTF-8 charset, HTTP response is encoded in some other charset).
When such an XML document is given without decoding/encoding-to-UTF8 to an XML parser that 
honors the charset directive, and attempts to interpret bytes as UTF-8, it will fail with 
some special characters that are not following the UTF-8 encoding (generally above the ANSI
range).

This is a patch that decodes all incoming HTTP responses with UTF-8, to match the claimed
character set in Dreambox's returned XML document. It works in the case of Dreambox but 
is obviously a wrong thing to do *if* any document would claim a character encoding other
than UTF-8!

!!! THEREFORE THIS PATCH SHOULD NEVER BE APPLIED TO RELEASE BRANCHES AS-IS !!!

Discussion is here: http://openremote.org/display/forums/XML+Readout+DREAMBOX?focusedCommentId=22872633#comment-22872633


+++ ../Controller_2_1_0_FM_DreamBoxPatch/src/org/openremote/controller/protocol/http/HttpGetCommand.java        (working copy)
@@ -263,7 +263,10 @@
           try
           {
             DocumentBuilder builder = factory.newDocumentBuilder();
-            ByteArrayInputStream bin = new ByteArrayInputStream(readValue.getBytes());
+
+            logger.warn("Replacing malformed input with UTF-8 default replacement byte array!");
+
+            ByteArrayInputStream bin = new ByteArrayInputStream(readValue.getBytes("UTF-8"));
             Document doc = builder.parse(bin);
 
             XPathFactory xfac = XPathFactory.newInstance();